This is madness!
I have diagnosed issues with HashMaps on plugins starts (on starts is
almost this same) and... it works but it not works. Methodes are
properly executing, HashMap have good values for... time of executing
method. If I use /st command, I get PlayerPoints from clear HashMap

But this is NOT the most strange thing. I have created command /st
HashMap where it... WORKS but structure of initiating method (arguments)
is probably this same.

Just madness...